* Fix PID corruption bug in group finding. * Add RVdoubleTag ReadAbacus Python interface. * Write multiplicity info to the groupstats log file. * Refactor the large FindAndProcessGlobalGroups() function to make better use of dependencies. * Change HaloStats to use RSD velocity convention.  Write header file in group/Step* directory. * Fix kick/unkick ordering for outputs and group finding. * Added new kick for GlobalGroup particle copies. * Timing breakdown for microstep.

+++ singlestep/DataModel/appendarena.cpp
@@ -348,7 +348,7 @@ class OutputRVdoubleTag: public AppendArena {
     struct ICparticle {
         double pos[3];		// Global position, unit box
 	    double vel[3];		// Zspace, unit box
-	    unsigned long tag;
+	    uint64 tag;
     };
 
     float velocity_conversion;
@@ -368,7 +368,7 @@ class OutputRVdoubleTag: public AppendArena {
 	p->vel[0] = vel.x*velocity_conversion;
 	p->vel[1] = vel.y*velocity_conversion;
 	p->vel[2] = vel.z*velocity_conversion;
-	p->tag = (unsigned int) aux.pid();
+	p->tag = aux.pid();
 	
     }
     void appendcell(char *c, integer3 ijk, float vscale) {
